feat: added stages to the action
