--- src/.eslintrc.custom.json
@@ -1,14 +1,0 @@
-{
-    "extends": "eslint:recommended",
-    "env": {
-        "browser": true,
-        "es2021": true
-    },
-    "parserOptions": {
-        "ecmaVersion": 12
-    },
-    "rules": {
-        "no-unused-vars": "warn",
-        "no-console": "off"
-    }
-}
